+++ src/main/kotlin/com/youquiz/quiz/QuizApplication.kt
@@ -1,11 +1,11 @@
-package com.youquiz.template
+package com.youquiz.quiz
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
-class TemplateApplication
+class QuizApplication
 
 fun main(args: Array<String>) {
-    runApplication<TemplateApplication>(*args)
+    runApplication<QuizApplication>(*args)
 }
