+++ src/test/kotlin/com/youquiz/quiz/QuizApplicationTest.kt
@@ -1,6 +1,6 @@
-package com.youquiz.template
+package com.youquiz.quiz
 
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class TemplateApplicationTest
+class QuizApplicationTest
