+++ src/test/kotlin/com/youquiz/quiz/config/TestConfiguration.kt
@@ -1,4 +1,4 @@
-package com.youquiz.template.config
+package com.youquiz.quiz.config
 
 import io.kotest.core.config.AbstractProjectConfig
 import io.kotest.core.extensions.Extension
