+++ src/test/kotlin/com/youquiz/quiz/util/TestUtil.kt
@@ -1,4 +1,4 @@
-package com.youquiz.template.util
+package com.youquiz.quiz.util
 
 import io.mockk.every
 import io.mockk.mockk
